+++ CHANGELOG.md
@@ -15,6 +15,13 @@
 - [Rust](https://github.com/moonrepo/plugins/blob/master/tools/rust/CHANGELOG.md)
 - [Schema (TOML, JSON, YAML)](https://github.com/moonrepo/plugins/blob/master/tools/internal-schema/CHANGELOG.md)
 
+## Unreleased
+
+#### 🚀 Updates
+
+- WASM API
+  - Added `InstallHook.forced` and `quiet` fields.
+
 ## 0.48.0
 
 #### 🚀 Updates